# openapi.yaml
openapi: 3.1.0
info:
  title: Writers Platform API
  description: A modern writing platform API built with FastAPI
  version: 1.0.0
paths:
...
  /api/v1/articles/:
    post:
      tags:
        - articles
      summary: Create Article
      description: |-
        Create a new article.

        Rate limit: 10 articles per minute per IP
        Requires: Valid JWT token, active user
      operationId: create_article_api_v1_articles__post
      security:
        - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - articles
      summary: Get Articles
      description: |-
        Get articles with filtering, sorting, and pagination.

        Rate limit: 30 requests per minute per IP
        Authentication: Optional
      operationId: get_articles_api_v1_articles__get
      security:
        - HTTPBearer: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            anyOf:
              - $ref: '#/components/schemas/ArticleStatus'
              - type: 'null'
            title: Status
        - name: category_id
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
              - type: 'null'
            title: Category Id
        - name: collection_id
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
              - type: 'null'
            title: Collection Id
        - name: author_id
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
              - type: 'null'
            title: Author Id
        - name: is_featured
          in: query
          required: false
          schema:
            anyOf:
              - type: boolean
              - type: 'null'
            title: Is Featured
        - name: tag
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Tag
        - name: search
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Search
        - name: skip
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
            title: Skip
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Limit
        - name: sort_by
          in: query
          required: false
          schema:
            type: string
            default: created_at
            title: Sort By
        - name: sort_order
          in: query
          required: false
          schema:
            type: string
            pattern: ^(asc|desc)$
            default: desc
            title: Sort Order
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedArticleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/collections/:
    post:
      tags:
        - collections
      summary: Create Collection
      description: |-
        Create a new collection (series/book).

        Rate limit: 5 collections per minute per IP
        Requires: Valid JWT token, active user
      operationId: create_collection_api_v1_collections__post
      security:
        - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - collections
      summary: Get Collections
      description: |-
        Get collections with filtering and pagination.

        Rate limit: 30 requests per minute per IP
        Authentication: Optional
      operationId: get_collections_api_v1_collections__get
      security:
        - HTTPBearer: []
      parameters:
        - name: type
          in: query
          required: false
          schema:
            anyOf:
              - $ref: '#/components/schemas/app__models__collection__CollectionType'
              - type: 'null'
            title: Type
        - name: status
          in: query
          required: false
          schema:
            anyOf:
              - $ref: '#/components/schemas/app__models__collection__CollectionStatus'
              - type: 'null'
            title: Status
        - name: author_id
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
              - type: 'null'
            title: Author Id
        - name: is_featured
          in: query
          required: false
          schema:
            anyOf:
              - type: boolean
              - type: 'null'
            title: Is Featured
        - name: skip
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
            title: Skip
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CollectionResponse'
                title: Response Get Collections Api V1 Collections  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/categories/:
    post:
      tags:
        - categories
      summary: Create Category
      description: |-
        Create a new category.
            
            Rate limit: 10 requests per minute per IP
            Requires: Admin privileges
      operationId: create_category_api_v1_categories__post
      security:
        - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - categories
      summary: Get Categories
      description: |-
        Get categories with filtering and pagination.
            
            Rate limit: 100 requests per minute per IP
            Public endpoint (no authentication required)
      operationId: get_categories_api_v1_categories__get
      security:
        - HTTPBearer: []
      parameters:
        - name: skip
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            description: Number of records to skip
            default: 0
            title: Skip
          description: Number of records to skip
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            description: Number of records to return
            default: 20
            title: Limit
          description: Number of records to return
        - name: parent_id
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
              - type: 'null'
            description: Filter by parent category
            title: Parent Id
          description: Filter by parent category
        - name: is_active
          in: query
          required: false
          schema:
            anyOf:
              - type: boolean
              - type: 'null'
            description: Filter by active status
            title: Is Active
          description: Filter by active status
        - name: search
          in: query
          required: false
          schema:
            anyOf:
              - type: string
                minLength: 1
                maxLength: 100
              - type: 'null'
            description: Search term
            title: Search
          description: Search term
        - name: include_children
          in: query
          required: false
          schema:
            type: boolean
            description: Include children in response
            default: false
            title: Include Children
          description: Include children in response
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'
                title: Response Get Categories Api V1 Categories  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ArticleCreate:
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 1
          title: Title
          description: Article title
        summary:
          anyOf:
            - type: string
              maxLength: 500
            - type: 'null'
          title: Summary
          description: Article summary
        content:
          type: string
          minLength: 1
          title: Content
          description: Article content
        meta_description:
          anyOf:
            - type: string
              maxLength: 160
            - type: 'null'
          title: Meta Description
          description: SEO meta description
        meta_keywords:
          anyOf:
            - type: string
              maxLength: 255
            - type: 'null'
          title: Meta Keywords
          description: SEO keywords
        category_id:
          anyOf:
            - type: integer
              exclusiveMinimum: 0
            - type: 'null'
          title: Category Id
          description: Category ID
        collection_id:
          anyOf:
            - type: integer
              exclusiveMinimum: 0
            - type: 'null'
          title: Collection Id
          description: Collection ID
        order_in_collection:
          anyOf:
            - type: integer
              exclusiveMinimum: 0
            - type: 'null'
          title: Order In Collection
          description: Order in collection
        allow_comments:
          type: boolean
          title: Allow Comments
          description: Allow comments on article
          default: true
        is_featured:
          type: boolean
          title: Is Featured
          description: Featured article
          default: false
        status:
          $ref: '#/components/schemas/ArticleStatus'
          description: Article status
          default: draft
        scheduled_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Scheduled At
          description: Scheduled publication time
        tag_names:
          anyOf:
            - items:
                type: string
              type: array
            - type: 'null'
          title: Tag Names
          description: List of tag names
      type: object
      required:
        - title
        - content
      title: ArticleCreate
      description: Schema for creating a new article.
...
    ArticleResponse:
      properties:
        id:
          type: integer
          title: Id
        title:
          type: string
          title: Title
        slug:
          type: string
          title: Slug
        summary:
          anyOf:
            - type: string
            - type: 'null'
          title: Summary
        content:
          type: string
          title: Content
        meta_description:
          anyOf:
            - type: string
            - type: 'null'
          title: Meta Description
        reading_time:
          anyOf:
            - type: integer
            - type: 'null'
          title: Reading Time
        status:
          $ref: '#/components/schemas/ArticleStatus'
        is_featured:
          type: boolean
          title: Is Featured
        allow_comments:
          type: boolean
          title: Allow Comments
        order_in_collection:
          anyOf:
            - type: integer
            - type: 'null'
          title: Order In Collection
        view_count:
          type: integer
          title: View Count
        like_count:
          type: integer
          title: Like Count
        comment_count:
          type: integer
          title: Comment Count
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Updated At
        published_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Published At
        author:
          $ref: '#/components/schemas/app__schemas__user__UserResponse'
        category:
          anyOf:
            - $ref: '#/components/schemas/CategoryResponse'
            - type: 'null'
        collection:
          anyOf:
            - $ref: '#/components/schemas/CollectionSummary'
            - type: 'null'
        tags:
          items:
            $ref: '#/components/schemas/TagResponse'
          type: array
          title: Tags
          default: []
      type: object
      required:
        - id
        - title
        - slug
        - content
        - status
        - is_featured
        - allow_comments
        - view_count
        - like_count
        - comment_count
        - created_at
        - author
      title: ArticleResponse
      description: Complete article response schema.
    ArticleStatus:
      type: string
      enum:
        - draft
        - published
        - archived
        - scheduled
      title: ArticleStatus
      description: Article publication status.
...
    CategoryCreate:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
          description: Category name
        description:
          anyOf:
            - type: string
              maxLength: 500
            - type: 'null'
          title: Description
          description: Category description
        color:
          anyOf:
            - type: string
              pattern: ^#[0-9A-Fa-f]{6}$
            - type: 'null'
          title: Color
          description: Category color (hex)
        icon:
          anyOf:
            - type: string
              maxLength: 50
            - type: 'null'
          title: Icon
          description: Category icon name
        parent_id:
          anyOf:
            - type: integer
            - type: 'null'
          title: Parent Id
          description: Parent category ID for hierarchical structure
      type: object
      required:
        - name
      title: CategoryCreate
      description: Schema for creating a new category.
...
    CategoryResponse:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
          description: Category name
        description:
          anyOf:
            - type: string
              maxLength: 500
            - type: 'null'
          title: Description
          description: Category description
        color:
          anyOf:
            - type: string
              pattern: ^#[0-9A-Fa-f]{6}$
            - type: 'null'
          title: Color
          description: Category color (hex)
        icon:
          anyOf:
            - type: string
              maxLength: 50
            - type: 'null'
          title: Icon
          description: Category icon name
        id:
          type: integer
          title: Id
        slug:
          type: string
          title: Slug
        parent_id:
          anyOf:
            - type: integer
            - type: 'null'
          title: Parent Id
        is_active:
          type: boolean
          title: Is Active
          default: true
        article_count:
          type: integer
          title: Article Count
          description: Number of articles in this category
          default: 0
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Updated At
      type: object
      required:
        - name
        - id
        - slug
        - created_at
      title: CategoryResponse
      description: Schema for category responses.
...
    CollectionCreate:
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 1
          title: Title
          description: Collection title
        description:
          anyOf:
            - type: string
              maxLength: 1000
            - type: 'null'
          title: Description
          description: Collection description
        type:
          $ref: '#/components/schemas/app__schemas__collection__CollectionType'
          description: Collection type (series or book)
        cover_image:
          anyOf:
            - type: string
            - type: 'null'
          title: Cover Image
          description: Cover image URL
      type: object
      required:
        - title
        - type
      title: CollectionCreate
      description: Schema for creating a new collection.
    CollectionResponse:
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 1
          title: Title
          description: Collection title
        description:
          anyOf:
            - type: string
              maxLength: 1000
            - type: 'null'
          title: Description
          description: Collection description
        type:
          $ref: '#/components/schemas/CollectionType-Output'
          description: Collection type (series or book)
        cover_image:
          anyOf:
            - type: string
            - type: 'null'
          title: Cover Image
          description: Cover image URL
        id:
          type: integer
          title: Id
        slug:
          type: string
          title: Slug
        status:
          $ref: '#/components/schemas/CollectionStatus-Output'
        author_id:
          type: integer
          title: Author Id
        article_count:
          type: integer
          title: Article Count
          description: Number of articles in collection
          default: 0
        total_views:
          type: integer
          title: Total Views
          description: Total views across all articles
          default: 0
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Updated At
        published_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Published At
      type: object
      required:
        - title
        - type
        - id
        - slug
        - status
        - author_id
        - created_at
      title: CollectionResponse
      description: Schema for collection responses.
    CollectionStatus-Output:
      type: string
      enum:
        - draft
        - published
        - archived
      title: CollectionStatus
      description: Collection status enumeration.
    CollectionSummary:
      properties:
        id:
          type: integer
          title: Id
        title:
          type: string
          title: Title
        slug:
          type: string
          title: Slug
        type:
          $ref: '#/components/schemas/CollectionType-Output'
        status:
          $ref: '#/components/schemas/CollectionStatus-Output'
        article_count:
          type: integer
          title: Article Count
          description: Number of articles in collection
          default: 0
      type: object
      required:
        - id
        - title
        - slug
        - type
        - status
      title: CollectionSummary
      description: Collection summary schema for article relations (no circular import).
    CollectionType-Output:
      type: string
      enum:
        - series
        - book
      title: CollectionType
      description: Collection type enumeration.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    PaginatedArticleResponse:
      properties:
        articles:
          items:
            $ref: '#/components/schemas/ArticleListResponse'
          type: array
          title: Articles
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
        total_pages:
          type: integer
          title: Total Pages
      type: object
      required:
        - articles
        - total
        - page
        - size
        - total_pages
      title: PaginatedArticleResponse
      description: Paginated article response.
...
    TagResponse:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        slug:
          type: string
          title: Slug
        usage_count:
          type: integer
          title: Usage Count
          description: Number of articles using this tag
          default: 0
        is_active:
          type: boolean
          title: Is Active
          default: true
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Updated At
      type: object
      required:
        - id
        - name
        - slug
        - created_at
      title: TagResponse
      description: Schema for tag responses.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
    app__models__collection__CollectionStatus:
      type: string
      enum:
        - draft
        - published
        - completed
        - archived
      title: CollectionStatus
      description: Collection publication status.
    app__models__collection__CollectionType:
      type: string
      enum:
        - series
        - book
        - anthology
        - course
      title: CollectionType
      description: Collection type enum.
...
    app__schemas__collection__CollectionType:
      type: string
      enum:
        - series
        - book
      title: CollectionType
      description: Collection type enumeration.
    app__schemas__user__UserResponse:
      properties:
        id:
          type: integer
          title: Id
        username:
          type: string
          title: Username
        full_name:
          anyOf:
            - type: string
            - type: 'null'
          title: Full Name
        avatar_url:
          anyOf:
            - type: string
            - type: 'null'
          title: Avatar Url
      type: object
      required:
        - id
        - username
      title: UserResponse
      description: Basic user response schema for article relations.
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
    HTTPBearer:
      type: http
      scheme: bearer